moved function into it's own utils folder
