+++ tests/unit/test_clean_cqc_location_data.py
@@ -890,180 +890,5 @@ class AddColumnRelatedLocation(CleanCQCLocationDatasetTests):
         )
 
 
-class ExtractRegisteredManagerNamesTests(CleanCQCLocationDatasetTests):
-    def setUp(self) -> None:
-        super().setUp()
-
-        self.extract_registered_manager_df = self.spark.createDataFrame(
-            Data.extract_registered_manager_rows,
-            Schemas.extract_registered_manager_schema,
-        )
-
-    @patch("jobs.clean_cqc_location_data.join_names_column_into_original_df")
-    @patch("jobs.clean_cqc_location_data.group_and_collect_names")
-    @patch("jobs.clean_cqc_location_data.select_and_create_full_name")
-    @patch("jobs.clean_cqc_location_data.extract_contacts_information")
-    def test_extract_registered_manager_names_calls_all_functions(
-        self,
-        extract_contacts_information_mock: Mock,
-        select_and_create_full_name_mock: Mock,
-        group_and_collect_names_mock: Mock,
-        join_names_column_into_original_df_mock: Mock,
-    ):
-        job.extract_registered_manager_names(self.extract_registered_manager_df)
-
-        extract_contacts_information_mock.assert_called_once()
-        select_and_create_full_name_mock.assert_called_once()
-        group_and_collect_names_mock.assert_called_once()
-        join_names_column_into_original_df_mock.assert_called_once()
-
-    def test_extract_registered_manager_names_returns_the_same_number_of_rows(self):
-        returned_df = job.extract_registered_manager_names(
-            self.extract_registered_manager_df
-        )
-        self.assertEqual(
-            self.extract_registered_manager_df.count(), returned_df.count()
-        )
-
-    def test_extract_contacts_information_returns_one_row_per_activity_with_a_contact(
-        self,
-    ):
-        test_df = self.spark.createDataFrame(
-            Data.extract_contacts_information_two_activities_one_contact_each_rows,
-            Schemas.extract_contacts_information_schema,
-        )
-        expected_df = self.spark.createDataFrame(
-            Data.expected_extract_contacts_information_two_activities_one_contact_each_rows,
-            Schemas.expected_extract_contacts_information_schema,
-        )
-        returned_df = job.extract_contacts_information(test_df)
-        self.assertEqual(
-            expected_df.sort(CQCL.location_id, CQCLCleaned.contacts_exploded).collect(),
-            returned_df.sort(CQCL.location_id, CQCLCleaned.contacts_exploded).collect(),
-        )
-
-    def test_extract_contacts_information_returns_one_row_per_contact_when_mutliple_contacts(
-        self,
-    ):
-        test_df = self.spark.createDataFrame(
-            Data.extract_contacts_information_one_activity_two_contacts_rows,
-            Schemas.extract_contacts_information_schema,
-        )
-        expected_df = self.spark.createDataFrame(
-            Data.expected_extract_contacts_information_one_activity_two_contacts_rows,
-            Schemas.expected_extract_contacts_information_schema,
-        )
-        returned_df = job.extract_contacts_information(test_df)
-        self.assertEqual(
-            expected_df.sort(CQCL.location_id, CQCLCleaned.contacts_exploded).collect(),
-            returned_df.sort(CQCL.location_id, CQCLCleaned.contacts_exploded).collect(),
-        )
-
-    def test_extract_contacts_information_doesnt_return_a_row_if_no_contacts(self):
-        test_df = self.spark.createDataFrame(
-            Data.extract_contacts_information_two_activities_but_only_one_with_contacts_rows,
-            Schemas.extract_contacts_information_schema,
-        )
-        expected_df = self.spark.createDataFrame(
-            Data.expected_extract_contacts_information_two_activities_but_only_one_with_contacts_rows,
-            Schemas.expected_extract_contacts_information_schema,
-        )
-        returned_df = job.extract_contacts_information(test_df)
-        self.assertEqual(
-            expected_df.collect(),
-            returned_df.collect(),
-        )
-
-    def test_select_and_create_full_name_returns_expected_columns(self):
-        test_df = self.spark.createDataFrame(
-            Data.select_and_create_full_name_rows,
-            Schemas.select_and_create_full_name_schema,
-        )
-        expected_df = self.spark.createDataFrame(
-            Data.expected_select_and_create_full_name_rows,
-            Schemas.expected_select_and_create_full_name_schema,
-        )
-        returned_df = job.select_and_create_full_name(test_df)
-        self.assertEqual(expected_df.columns, returned_df.columns)
-
-    def test_select_and_create_full_name_returns_expected_rows(self):
-        test_df = self.spark.createDataFrame(
-            Data.select_and_create_full_name_rows,
-            Schemas.select_and_create_full_name_schema,
-        )
-        expected_df = self.spark.createDataFrame(
-            Data.expected_select_and_create_full_name_rows,
-            Schemas.expected_select_and_create_full_name_schema,
-        )
-        returned_df = job.select_and_create_full_name(test_df)
-        self.assertEqual(
-            expected_df.sort(CQCL.location_id).collect(),
-            returned_df.sort(CQCL.location_id).collect(),
-        )
-
-    def test_group_and_collect_names_removes_duplicate_names(self):
-        test_df = self.spark.createDataFrame(
-            Data.group_and_collect_names_with_duplicate_names_rows,
-            Schemas.group_and_collect_names_schema,
-        )
-        expected_df = self.spark.createDataFrame(
-            Data.expected_group_and_collect_names_with_duplicate_names_rows,
-            Schemas.expected_group_and_collect_names_schema,
-        )
-        returned_df = job.group_and_collect_names(test_df)
-        self.assertEqual(
-            expected_df.collect(),
-            returned_df.collect(),
-        )
-
-    def test_group_and_collect_names_groups_by_locationid_and_import_date_correctly(
-        self,
-    ):
-        test_df = self.spark.createDataFrame(
-            Data.group_and_collect_names_with_different_ids_and_dates_rows,
-            Schemas.group_and_collect_names_schema,
-        )
-        expected_df = self.spark.createDataFrame(
-            Data.expected_group_and_collect_names_with_different_ids_and_dates_rows,
-            Schemas.expected_group_and_collect_names_schema,
-        )
-        returned_df = job.group_and_collect_names(test_df)
-        self.assertEqual(
-            expected_df.sort(
-                CQCL.location_id, CQCLCleaned.cqc_location_import_date
-            ).collect(),
-            returned_df.sort(
-                CQCL.location_id, CQCLCleaned.cqc_location_import_date
-            ).collect(),
-        )
-
-    def test_join_names_column_into_original_df_returns_expected_df(
-        self,
-    ):
-        original_df = self.spark.createDataFrame(
-            Data.original_test_rows,
-            Schemas.original_test_schema,
-        )
-        registered_manager_names_df = self.spark.createDataFrame(
-            Data.registered_manager_names_rows,
-            Schemas.registered_manager_names_schema,
-        )
-        expected_joined_df = self.spark.createDataFrame(
-            Data.expected_join_with_original_rows,
-            Schemas.expected_join_with_original_schema,
-        )
-        returned_joined_df = job.join_names_column_into_original_df(
-            original_df, registered_manager_names_df
-        )
-        self.assertEqual(
-            expected_joined_df.sort(
-                CQCL.location_id, CQCLCleaned.cqc_location_import_date
-            ).collect(),
-            returned_joined_df.sort(
-                CQCL.location_id, CQCLCleaned.cqc_location_import_date
-            ).collect(),
-        )
-
-
 if __name__ == "__main__":
     unittest.main(warnings="ignore")
